added data generator

+++ server/scheduler123/src/main/java/org/scheduler/Flatbed.java
@@ -1,7 +1,7 @@
 package org.scheduler;
 
 public class Flatbed extends Trailer{
-    public Flatbed(int load, int loadCapacity, Carrier carrier, int timePerLbs) {
+    public Flatbed(int load, int loadCapacity, Carrier carrier, double timePerLbs) {
         super(load, loadCapacity, carrier, timePerLbs);
     }
     // A trailer that consists of only a deck, uses tarp to cover goods
