+++ server/scheduler123/src/main/java/org/scheduler/Van.java
@@ -2,7 +2,7 @@ package org.scheduler;
 
 public class Van extends Trailer{
 
-    public Van(int load, int loadCapacity, Carrier carrier, int timePerLbs) {
+    public Van(int load, int loadCapacity, Carrier carrier, double timePerLbs) {
         super(load, loadCapacity, carrier, timePerLbs);
     }
 }
